Modify the default PatchDock parameter file.
We want to use non-default parameters for the PatchDock
docking. Previously, CryptoSite used a customized copy
of PatchDock with a modified buildParams.pl script to
do this. Obviously this won't work outside of the Sali
lab, so instead use the stock PatchDock binaries and
modify the parameter file after it's generated.